requirements.txt
lemon==1.0.3
paho-mqtt==1.6.1
asyncio-mqtt==0.12.1
requests==2.28.1
python-dotenv==0.20.0
